Remove duplicate schemas
